reorganized structure and bug fixes
