+++ api/routers/users.py
@@ -1,0 +1,103 @@
+from fastapi import APIRouter
+
+from datetime import datetime, timedelta
+from enum import Enum
+from fastapi import Depends, FastAPI, HTTPException, status, Request, Form
+from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+from jose import JWTError, jwt
+from passlib.context import CryptContext
+from pydantic import BaseModel
+from database import users_table
+from database import mail_verif_table
+from validation import *
+
+from ..dependencies.authentication import create_access_token, Token, get_password_hash, ACCESS_TOKEN_EXPIRE_MINUTES
+from ..dependencies.users import get_current_user, get_user, User, UserWithSensitiveInfo, authenticate_user
+
+router = APIRouter()
+
+
+
+@router.get("/users/me/", response_model=User)
+async def read_users_me(current_user: User = Depends(get_current_user)):
+    """API call to get the curret user we are communicating with
+
+    Args:
+        current_user (User, optional): User. Defaults to Depends(get_current_user).
+
+    Returns:
+        User: Current user with only the basic infos (no password)
+    """
+    return current_user
+
+@router.post("/users/login/", response_model=Token)
+async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
+    """API call to create an access token
+
+    Args:
+        form_data (OAuth2PasswordRequestForm, optional): Login data. Defaults to Depends().
+
+    Raises:
+        HTTPException: Raises error if the email or password or incorrect
+
+    Returns:
+        dict: Token information in json format
+    """
+    
+    #note: username is the reserved name for the login name, must be used even if we are using email
+    user = authenticate_user(form_data.username, form_data.password)
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect username or password",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    access_token = create_access_token(
+        data={"sub": user.email}, expires_delta=access_token_expires
+    )
+    return {"access_token": access_token, "token_type": "bearer"}
+
+@router.post("/users/signup/", status_code=status.HTTP_201_CREATED)
+async def register(email: str = Form(), password: str = Form(), first_name: str = Form(), last_name: str = Form()):
+    """API call to create a new account
+
+    Args:
+        email (str, optional): Email. Defaults to Form().
+        password (str, optional): Plaintext password. Defaults to Form().
+        first_name (str, optional): Frist name. Defaults to Form().
+        last_name (str, optional): Last name Defaults to Form().
+
+    Raises:
+        HTTPException: _description_
+        HTTPException: _description_
+
+    Returns:
+        _type_: _description_
+    """
+    errors = []
+    validate_email(email)
+    errors.extend(validate_password(password))
+    errors.extend(validate_first_name(first_name))
+    errors.extend(validate_last_name(last_name))
+    if len(errors) > 0:
+        raise HTTPException(
+            status_code=status.HTTP_406_NOT_ACCEPTABLE,
+            detail=errors,
+        )
+    if get_user(email):
+        raise HTTPException(
+            status_code=status.HTTP_406_NOT_ACCEPTABLE,
+            detail="This email is already assosiated with an existing account",
+        )
+    
+    hashed_pw = get_password_hash(password)
+    user = UserWithSensitiveInfo(   email=email,
+                                    first_name=first_name,
+                                    last_name=last_name,
+                                    hashed_password=hashed_pw,
+                                    permission=1,
+                                    disabled=0,
+                                    email_verified=0)
+    users_table.create_user(user)
+    return {"message": "success"}
