+++ tests/tests.py
@@ -1,0 +1,1 @@
+#run all tests here
